split authentication and bomnmodeler
