+++ src/contracts/authentication/IAuthenticationRepository.ts
@@ -1,0 +1,8 @@
+import {ILoginResult} from './ILoginResult';
+import {ILogoutResult} from './ILogoutResult';
+
+export interface IAuthenticationRepository {
+  login(username: string, password: string): Promise<ILoginResult>;
+  logout(): Promise<ILogoutResult>;
+  getIdentity(token: string): Promise<any>;
+}
