+++ src/contracts/authentication/IAuthenticationService.ts
@@ -1,0 +1,9 @@
+import {IIdentity} from './IIdentity';
+
+export interface IAuthenticationService {
+  login(username: string, password: string): Promise<IIdentity>;
+  logout(): Promise<void>;
+  getToken(): string;
+  hasToken(): boolean;
+  getIdentity(): IIdentity;
+}
