+++ src/contracts/authentication/IIdentity.ts
@@ -1,0 +1,5 @@
+export interface IIdentity {
+  id: string;
+  name: string;
+  roles: Array<string>;
+}
