+++ src/contracts/authentication/ILoginResult.ts
@@ -1,0 +1,6 @@
+import {IIdentity} from './IIdentity';
+
+export interface ILoginResult {
+  identity: IIdentity;
+  token: string;
+}
