+++ src/contracts/authentication/ILogoutResult.ts
@@ -1,0 +1,3 @@
+export interface ILogoutResult {
+  result: boolean;
+}
