+++ src/contracts/authentication/constants.ts
@@ -1,0 +1,4 @@
+export enum AuthenticationStateEvent {
+  LOGIN = 'login',
+  LOGOUT = 'logout',
+}
