+++ src/contracts/authentication/index.ts
@@ -1,0 +1,6 @@
+export * from './constants';
+export {IAuthenticationRepository} from './IAuthenticationRepository';
+export {IAuthenticationService} from './IAuthenticationService';
+export {IIdentity} from './IIdentity';
+export {ILoginResult} from './ILoginResult';
+export {ILogoutResult} from './ILogoutResult';
