+++ src/contracts/bpmnmodeler/IBpmnModeler.ts
@@ -1,0 +1,8 @@
+export interface IBpmnModeler {
+  definitions: any;
+  attachTo(wrapper: HTMLElement): void;
+  saveXML(options: any,
+          callback: (error: Error, result: String) => void): void;
+  importXML(xml: string,
+            errorHandler: (err: Error) => void): void;
+}
