+++ src/contracts/bpmnmodeler/IBpmnModelerConstructor.ts
@@ -1,0 +1,9 @@
+import {IBpmnModeler} from './IBpmnModeler';
+import {IDependencyHook} from './IDependencyHook';
+
+export interface IBpmnModelerConstructor {
+  new(options: {
+    additionalModules?: Array<IDependencyHook>,
+    container?: string,
+  }): IBpmnModeler;
+}
