+++ src/contracts/bpmnmodeler/ICanvas.ts
@@ -1,0 +1,5 @@
+import {IShape} from './IShape';
+
+export interface ICanvas {
+  getRootElement(): IShape;
+}
