+++ src/contracts/bpmnmodeler/IDependencyHook.ts
@@ -1,0 +1,5 @@
+export interface IDependencyHook {
+  __depends__: Array<string>;
+  __init__: Array<string>;
+  [index: string]: [string, any] | Array<string>;
+}
