+++ src/contracts/bpmnmodeler/IModdleElement.ts
@@ -1,0 +1,7 @@
+export interface IModdleElement {
+  id: string;
+  get: any;
+  $type: string;
+  $attrs?: any;
+  $parent?: IModdleElement;
+}
