+++ src/contracts/bpmnmodeler/IShape.ts
@@ -1,0 +1,12 @@
+import {IModdleElement} from './IModdleElement';
+
+export interface IShape {
+  businessObject: IModdleElement;
+  id: string;
+  type: string;
+  label: IShape;
+  x: number;
+  y: number;
+  width: number;
+  height: number;
+}
