+++ src/contracts/bpmnmodeler/index.ts
@@ -1,0 +1,6 @@
+export {IBpmnModeler} from './IBpmnModeler';
+export {IBpmnModelerConstructor} from './IBpmnModelerConstructor';
+export {ICanvas} from './ICanvas';
+export {IDependencyHook} from './IDependencyHook';
+export {IModdleElement} from './IModdleElement';
+export {IShape} from './IShape';
