+++ src/contracts/index.ts
@@ -1,6 +1,6 @@
 export * from './iprocessengine';
-export * from './ibpmnmodeler';
+export * from './bpmnmodeler/index';
 export * from './idynamic-ui';
 export * from './ichoose-dialog';
-export * from './iauthentication';
+export * from './authentication/index';
 export * from './ieventbus';
